==> src/3_descriptive_stats_table.ipynb <==
import os
import pandas as pd
import numpy as np
from sspipe import p

# set up paths
origin_path = os.path.abspath('..') | p(os.path.join, 'data', 'tab')
derived_path = os.path.abspath('..') | p(os.path.join, 'data', 'derived')

# read data
sample = pd.read_csv(os.path.join(derived_path, 'sample_cleaned.csv'), low_memory=False)

# descriptive variables
desc_vars = ['sex_1', 'age_2002', 'married_1', 'grandchild_1', 'outside_uk_1',
             'degree_1', 'below_degree_1', 'a_levels_1', 'o_levels_1', 'no_qual_1',
             'job_years_1', 'job_permanent_1', 'job_30h_1', 'reach_spa_2004', 'early_retire_incentive_1',
             'pen_db_1', 'pen_dc_1', 'pen_private_1', 'total_pension_2002', 'total_wealth_1', 'empinc_r_s_1',
             'adl_1', 'poor_health_1', 'cardio_1', 'noncardio_1', 'ex_alive_1', 'ex_work_1', 'ex_limit_1']

desc_df = sample.groupby(['treatment'])[desc_vars].mean().round(2).T
desc_df.rename(columns={1: 'Retired', 0: 'Employed'}, inplace=True)
desc_df = desc_df[['Employed', 'Retired']]
desc_df.index = ['Female', 'Age in year 2002', 'Married or cohabiting', 'Has grandchildren', 'Country of birth outside the UK',
                 'University degree', 'Higher education below degree', 'A levels', 'O levels', 'No qualification',
                 'Number of years worked in current job', 'Current job is permanent', '1-30h work per week at current job',
                 'Will reach state pension age at wave 2 interview', 'Has been offered early retirement incentives',
                 'Has defined benefit pension scheme', 'Has defined contribution pension scheme',
                 'Has any private pension scheme', 'Total pension wealth in 2002', 'Net (non-pension) total wealth',
                 'Employment income per week', 'Number of difficulties in ADL', 'Self-assessed health poor or fair',
                 'Has diagnosed cardiovascular diseases', 'Has diagnosed non-cardiovascular diseases',
                 'Likelihood of being alive', 'Likelihood of working', 'Likelihood that health limits ability to work']

print(desc_df.to_latex(float_format="%.2f"))
