==> src/10_health_after_retire_table_ext.ipynb <==
import os
import pandas as pd
import numpy as np
import statsmodels.formula.api as smf
import plotnine as pn
from sspipe import p, px

# Set up paths
derived_path = os.path.abspath('..') | p(os.path.join, 'data', 'derived')

# Read data
sample = pd.read_csv(os.path.join(derived_path, 'sample_simulate_retire_ext.csv'))

# Any newly diagnosed disease
disease_list = ['angina', 'heart_attack', 'diabetes', 'stroke', 'arthritis', 'cancer', 'psych']
sample['any_post'] = (sample[[disease + '_post' for disease in disease_list]] == 1).any(axis=1)

vars = ['any_post'] + [disease + '_post' for disease in disease_list] + ['poor_health_3', 'limit_3', 'adl_3']

df = sample.groupby(['treatment'])[vars].mean().round(3).T
df.rename(columns={1: 'Retired', 0: 'Employed'}, inplace=True)
df = df[['Employed', 'Retired']]
df.index = ['Any newly diagnosed disease', 
            'Newly diagnosed angina', 'New heart attack', 'New stroke', 'Newly diagnosed diabetes',
            'Newly diagnosed arthritis', 'Newly diagnosed cancer', 'Newly diagnosed psychiatric problem',
            'Poor self-assessed health', 'Limiting long-standing illness', 'Number of difficulties in ADL']

print(df.to_latex(float_format="%.2f"))

**Difference**
Numbers for retired individuals should be identical, since their retirement dates are not imputed.

Different items:
- Newly diagnosed angina
- New heart attack (not much)
- New stroke (not much)
- Newly diagnosed arthritis (not much)
- Newly diagnosed cancer (not much)
- Newly diagnosed psych (not much)
- Limiting long-standing illness
- Number of difficulties in ADL
