==> src/13_main_table_ext.ipynb <==
import os
import pandas as pd
import numpy as np
from sspipe import p, px

# Set up paths
derived_path = os.path.abspath('..') | p(os.path.join, 'data', 'derived')
# Read data
table_data = pd.read_csv(os.path.join(derived_path, 'main_table_ext.csv'))

# Round
main_table = table_data.apply(lambda row: round(row, 3) if (row.name + 1) % 2 == 1 else round(row, 2),
                              axis=1)

# Add * to indicate statistical significance
for i in np.arange(0, 18, 2):
    for j in range(4):
        if abs(main_table.iloc[i + 1, j]) >= 3.29:
            main_table.iloc[i, j] = str(main_table.iloc[i, j]) + '***'
        elif abs(main_table.iloc[i + 1, j]) >= 1.96:
            main_table.iloc[i, j] = str(main_table.iloc[i, j]) + '**'

# Add square brackets to t-values
main_table = main_table.apply(lambda row: '[' + row.astype(str) + ']' if row.name % 2 == 1 else row,
                              axis=1)

# Change index
main_table.index = ['Any chronic disease', '', 'Angina, heart attack or stroke', '', 'Angina', '', 'Heart attack', '', 'Stroke',
                    '', 'Diabetes', '', 'Arthritis', '', 'Cancer', '', 'Psychiatric', '']
# Change column names
main_table.rename(columns={'Xset1': 'ATET\nXset 1', 'Xset2': 'ATET\nXset 2', 'Xset3': 'ATET\nXset 3', 'Xset4': 'ATET\nXset 4'},
                  inplace=True)

# Get LaTeX code
print(main_table.to_latex(float_format="%.3f"))
